Refactor for Bazel integration test runner
