+++ BUILD.bazel
@@ -81,17 +81,23 @@ rust_test(
         ":evm_rpc_debug.wasm",
         "//rs/canister_sandbox",
         "//rs/canister_sandbox/sandbox_launcher",
+        "//rs/rosetta-api/icrc1/ledger:ledger_canister_u256.wasm",
     ],
     env = {
         "CARGO_MANIFEST_DIR": "rs/ethereum/evm_rpc",
         "EVM_RPC_WASM_PATH": "$(rootpath :evm_rpc_debug.wasm)",
+        "IC_ICRC1_LEDGER_WASM_PATH": "$(rootpath //rs/rosetta-api/icrc1/ledger:ledger_canister_u256.wasm)",
         "LAUNCHER_BINARY": "$(rootpath //rs/canister_sandbox/sandbox_launcher)",
         "SANDBOX_BINARY": "$(rootpath //rs/canister_sandbox)",
     },
-    proc_macro_deps = [],
+    proc_macro_deps = [
+        "@crate_index//:strum_macros",
+    ],
     deps = [
         ":evm_rpc_lib",
+        "//packages/icrc-ledger-types:icrc_ledger_types",
         "//rs/ethereum/cketh/minter:minter",
+        "//rs/rosetta-api/icrc1/ledger",
         "//rs/rust_canisters/http_types",
         "//rs/state_machine_tests",
         "//rs/test_utilities/load_wasm",
